add price filtration for all products only

+++ app/src/main/java/com/example/yallabuy_user/products/ProductsViewModelInterface.kt
@@ -5,5 +5,5 @@ interface ProductsViewModelInterface {
     suspend fun getAllCategories()
     suspend fun getProducts(categoryID: Long?)
     fun showSubCategoryProduct(subCategory : String)
-    fun showFilteredProduct(minPrice : Double, maxPrice :Double)
+    fun showFilteredProduct(minPrice : Float, maxPrice :Float)
 }
